requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
requests==2.32.3


